Setup of website
This is a small but crucial step to begin building the website. In this setup stage I have started cleaning up the old website by removing all the old files and images in a seperate folder. This makes it easy as I can access tis folder at any time in order to look back at code. Secondly, I have created all the page names that I will be using for my website so I know which page is which.