Setting up layout
I have had progress and have begun setting up my dropdown icons. This is an important step as the dropdown menus are the most important forms of navigation. I am now trying to get this step right making sure it is the right size and functions exactly how I want it to.