Started adding information
I have added all of the information to the Content 1 page. This includes the immediate impact that New Zealand will see in the coming days. This completes the main portion of information that I have to do and now I have to move on to the secondary information on my topic.